requirements.txt
openai>=1.76.2
pyyaml>=6.0.2
python-dotenv>=1.0.1
jsonschema>=4.23.0
docopt-ng>=0.9.0
langchain-community>=0.3.16
duckduckgo-search>=7.3.0
streamlit>=1.44.0
streamlit-mermaid>=0.3.0
psutil>=7.0.0
openapi>=2.0.0
openai-agents[litellm]>=0.0.14
pycron>=3.1.2
beeai-framework>=0.1.17
slack_sdk>=3.35.0
nest-asyncio>=1.6.0
pydantic-ai[logfire]>=0.1.8
opik>=1.7.22

# Dev dependencies
black>=24.10.0

# Test dependencies
pytest>=8.3.4
pytest-mock>=3.14.0 